Fix queue/0 type deprecations (with queue:queue/0)
When including the latest on `basho:develop` for poolboy as a dependency in
an Erlang project, rebar aborted with a warning as error with the following
messages:

```
/Users/susan.potter/Code/oss/carson/deps/poolboy/src/poolboy.erl:24: type queue/0 is deprecated and will be removed in
OTP 18.0; use use queue:queue/0 or preferably queue:queue/1
/Users/susan.potter/Code/oss/carson/deps/poolboy/src/poolboy.erl:25: type queue/0 is deprecated and will be removed in
OTP 18.0; use use queue:queue/0 or preferably queue:queue/1
ERROR: compile failed while processing /Users/susan.potter/Code/oss/carson/deps/poolboy: rebar_abort
```

*This pull request addresses the above deprecations so I can build keeping
with `warnings_as_errors` on, which I prefer. This might break compatibility
with much older versions of Erlang but for practical purposes I do not believe
this will pose a problem for versions that poolboy wishes to support.*

This is the [erlang-questions ML thread] [1] where Andreas Schumacher
outlines the deprecation warnings in R17 to prepare the way for R18.0.

[1]: http://erlang.org/pipermail/erlang-questions/2014-March/078081.html "R18.0 type deprecations"